requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=1.5.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
mediapipe>=0.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=0.21.0
Pillow>=9.5.0
